Move TrainerInitializationException to Trainer class.
