+++ tests/training_tests/test_training.py
@@ -1,10 +1,9 @@
 from tests.base_case import ChatBotTestCase
-from chatterbot import ChatBot
 from chatterbot.trainers import ListTrainer
 
 
 class TrainingTests(ChatBotTestCase):
 
     def test_trainer_not_set(self):
-        with self.assertRaises(ChatBot.TrainerInitializationException):
+        with self.assertRaises(ListTrainer.TrainerInitializationException):
             self.chatbot.train()
